Proyecto_Flask/requirements.txt
bcrypt==4.0.1
blinker==1.5
click==8.1.3
colorama==0.4.6        
dnspython==2.3.0       
email-validator==1.3.1 
Flask==2.2.2
Flask-Bcrypt==1.0.1    
Flask-Login==0.6.2     
Flask-Mail==0.9.1      
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==0.24
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
pycparser==2.18
six==1.11.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1